+++ src/index.ts
@@ -1,10 +1,8 @@
 export { createServiceCollection } from './createServiceCollection';
 export { dependsOn } from './dependsOn';
-export { IDisposable, IServiceCollection, IServiceModule, IServiceProvider, IServiceScope } from './interfaces';
-export type { AbstractNewable, InstanceFactory, LifetimeBuilder, MetadataType, Newable, ServiceBuilder, ServiceDescriptor, ServiceDescriptorConcrete, ServiceDescriptorFactory, ServiceIdentifier, ServiceImplementation, ServiceModuleType, SourceType, ServiceCollectionOptions } from './types';
-export { ResolveMultipleMode, DefaultServiceCollectionOptions, LogLevel } from './types';
+export { IDisposable, IServiceCollection, IServiceModule, IServiceProvider, IResolutionScope, IScopedProvider } from './interfaces';
+export type { AbstractNewable, InstanceFactory, ILifetimeBuilder, MetadataType, Newable, IServiceBuilder, ServiceDescriptor, ServiceIdentifier, ServiceImplementation, ServiceModuleType, SourceType, ServiceCollectionOptions } from './types';
+export { DefaultServiceCollectionOptions } from './defaults';
 export { ILogger } from './logger';
-export { MultipleRegistrationError, SelfDependencyError, ServiceCreationError, UnregisteredServiceError, ServiceError } from './errors';
-
-export { ServiceCollection } from './ServiceCollection';
-export { ServiceProvider } from './ServiceProvider';
+export { MultipleRegistrationError, SelfDependencyError, ServiceCreationError, UnregisteredServiceError, ServiceError, ScopedSingletonRegistrationError } from './errors';
+export { ResolveMultipleMode, Lifetime, LogLevel } from './enums';