+++ src/types.ts
@@ -1,7 +1,7 @@
-import type { ConsoleLogger } from './consoleLogger';
-import type { Lifetime } from './constants';
-import type { IServiceModule, IServiceProvider, IServiceScope } from './interfaces';
+import type { Lifetime, LogLevel, ResolveMultipleMode } from './enums';
+import type { IResolutionScope, IServiceModule } from './interfaces';
 import type { ILogger } from './logger';
+import type { ConsoleLogger } from './private/consoleLogger';
 
 export type SourceType = object;
 
@@ -11,63 +11,17 @@ export type Newable<T> = new (...args: any[]) => T;
 export type ServiceIdentifier<T extends SourceType> = { prototype: T; name: string }; //AbstractNewable<T>;
 export type ServiceImplementation<T extends SourceType> = Newable<T>;
 
-export type InstanceFactory<T extends SourceType> = (x: IServiceScope & IServiceProvider) => T;
+export type InstanceFactory<T extends SourceType> = (x: IResolutionScope) => T;
 
 export type ServiceModuleType = Newable<IServiceModule>;
 
-export type ServiceDescriptorFactory<T extends SourceType> = {
-  implementation: ServiceIdentifier<T>;
-  factory: InstanceFactory<T>;
+export interface ServiceDescriptor<T extends SourceType> {
+  readonly implementation: ServiceImplementation<T>;
   lifetime: Lifetime;
-};
-export type ServiceDescriptorConcrete<T extends SourceType> = {
-  implementation: ServiceImplementation<T>;
-  lifetime: Lifetime;
-};
-
-export type ServiceDescriptor<T extends SourceType> = ServiceDescriptorConcrete<T> | ServiceDescriptorFactory<T>;
-
-export type LifetimeBuilder = {
-  singleton: () => LifetimeBuilder;
-  scoped: () => LifetimeBuilder;
-  transient: () => LifetimeBuilder;
-};
-export type ServiceBuilder<T extends SourceType> = {
-  to: {
-    (implementation: ServiceImplementation<T>): LifetimeBuilder;
-    (implementation: ServiceIdentifier<T>, factory: InstanceFactory<T>): LifetimeBuilder;
-  };
-};
-
-export type MetadataType<T extends SourceType> = Record<string | symbol, ServiceIdentifier<T>>;
-
-export const ResolveMultipleMode = {
-  Error: 'ERROR',
-  LastRegistered: 'LAST_REGISTERED',
-} as const;
-export type ResolveMultipleMode = (typeof ResolveMultipleMode)[keyof typeof ResolveMultipleMode];
-
-export enum LogLevel {
-  Debug = 0,
-  Info = 1,
-  Warn = 2,
-  Error = 3,
-  None = 4,
+  createInstance(context: IResolutionScope): T;
 }
 
-// export const LogLevel = {
-//   Debug: 0,
-//   Info: 1,
-//   Warn: 2,
-//   Error: 3,
-//   None: 4,
-// } as const;
-// export type LogLevel = (typeof LogLevel)[keyof typeof LogLevel];
-
-export const DefaultServiceCollectionOptions: ServiceCollectionOptions = {
-  registrationMode: ResolveMultipleMode.Error,
-  logLevel: LogLevel.Warn,
-};
+export type MetadataType<T extends SourceType> = Record<string | symbol, ServiceIdentifier<T>>;
 
 export type ServiceCollectionOptions = {
   /**
@@ -87,3 +41,20 @@ export type ServiceCollectionOptions = {
    */
   logger?: ILogger;
 };
+
+export type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (k: infer I) => void ? I : never;
+
+export type EnsureObject<T> = T extends object ? T : never;
+
+export interface ILifetimeBuilder {
+  singleton(): ILifetimeBuilder;
+  scoped(): ILifetimeBuilder;
+  transient(): ILifetimeBuilder;
+}
+
+export interface IServiceBuilder<T extends SourceType> {
+  to: {
+    (implementation: ServiceImplementation<T>): ILifetimeBuilder;
+    (implementation: ServiceImplementation<T>, factory: InstanceFactory<T>): ILifetimeBuilder;
+  };
+}