+++ test/singleton.spec.ts
@@ -1,5 +1,5 @@
 import { equal, ok } from 'node:assert/strict';
-import { describe, it } from 'mocha';
+import { describe, it } from 'vitest';
 import { type IDisposable, createServiceCollection } from '../src';
 
 abstract class IAbstract {